fix: tidder takes $1 input, add user-agent lynx
Before tidder/reddit was not working because within a function $1 is
it's own thing. And whatever gets assigned to it gets reseted after
the function call.

So because it's it's own thing and not the same as
the shell parameter $1, the local variable is not $2. $2 is from the
shell input as in `tidder month linux`. So if you want to store 'linux' in
a global variable you would need $2. Or if you would have a function
call hardcoded in your script as in `month $2`.

But again inside a function those parameters are a different set. And
they ofcourse start with $1. My guess is that in a function $0 would be
the function name. As $0 for shell input is the script name.

* add user-agent to mask lynx