✨ add rgt script to search for #tags in files
This scripts uses ripgrep to search for content in files. Specifically it is
build to search for tags in files. The idea is to be able to search for
for multiple tags. And then only show results that contain all of the
searched tags.

If you want to search for multiple tags but you don't care if only one
of the searched tags show up use `rg '#foo|#foo2'`

TODO: implement the multi tag search with min 1 tag output.